Load navigation fixtures through common routes
