+++ playwright/settings-screenshot.spec.js
@@ -1,5 +1,4 @@
 import { test, expect } from "./fixtures/commonSetup.js";
-import fs from "fs";
 
 const runScreenshots = process.env.SKIP_SCREENSHOTS !== "true";
 
@@ -13,18 +12,6 @@ test.describe.parallel(
 
     for (const mode of modes) {
       test(`mode ${mode} collapsed`, async ({ page }) => {
-        await page.route("**/src/data/navigationItems.json", (route) => {
-          route.fulfill({ path: "tests/fixtures/navigationItems.json" });
-        });
-        await page.route("**/src/data/*.json", (route) => {
-          const file = route.request().url().split("/").pop();
-          const fixturePath = `tests/fixtures/${file}`;
-          if (fs.existsSync(fixturePath)) {
-            route.fulfill({ path: fixturePath });
-          } else {
-            route.continue();
-          }
-        });
         await page.addInitScript((mode) => {
           localStorage.setItem(
             "settings",
@@ -36,18 +23,6 @@ test.describe.parallel(
       });
 
       test(`mode ${mode} expanded`, async ({ page }) => {
-        await page.route("**/src/data/navigationItems.json", (route) => {
-          route.fulfill({ path: "tests/fixtures/navigationItems.json" });
-        });
-        await page.route("**/src/data/*.json", (route) => {
-          const file = route.request().url().split("/").pop();
-          const fixturePath = `tests/fixtures/${file}`;
-          if (fs.existsSync(fixturePath)) {
-            route.fulfill({ path: fixturePath });
-          } else {
-            route.continue();
-          }
-        });
         await page.addInitScript((mode) => {
           localStorage.setItem(
             "settings",
